+++ managementApp/src/main/java/managementapp/service/impl/MenuServiceImpl.java
@@ -25,7 +25,7 @@ public class MenuServiceImpl implements MenuService {
 	}
 
 	@Override
-	public List<Menu> findByPriceBelowLimit(Long priceLimit) {
+	public List<Menu> findByPriceBelowLimit(int priceLimit) {
 		return menuRepository.findByPriceBelow(priceLimit);
 	}
